Simplified all files
