+++ src/android/AndroidManifest.icons.json
