+++ src/ios/AppIcon.iconset.Contents.template.json
