+++ test/CordovaApp/platforms/ios/ionic_app/Images.xcassets/AppIcon.appiconset/Contents.json
@@ -21,12 +21,6 @@
     {
       "size": "60x60",
       "idiom": "iphone",
-      "scale": "3x",
-      "filename": "iphone-60x60-3x.png"
-    },
-    {
-      "size": "60x60",
-      "idiom": "iphone",
       "scale": "2x",
       "filename": "iphone-60x60-2x.png"
     },
@@ -67,28 +61,34 @@
       "filename": "iphone-29x29-1x.png"
     },
     {
-      "size": "50x50",
-      "idiom": "ipad",
-      "scale": "1x",
-      "filename": "ipad-50x50-1x.png"
-    },
-    {
       "size": "40x40",
       "idiom": "ipad",
       "scale": "2x",
       "filename": "ipad-40x40-2x.png"
     },
     {
-      "size": "72x72",
+      "size": "50x50",
       "idiom": "ipad",
       "scale": "1x",
-      "filename": "ipad-72x72-1x.png"
+      "filename": "ipad-50x50-1x.png"
     },
     {
       "size": "50x50",
       "idiom": "ipad",
       "scale": "2x",
       "filename": "ipad-50x50-2x.png"
+    },
+    {
+      "size": "60x60",
+      "idiom": "iphone",
+      "scale": "3x",
+      "filename": "iphone-60x60-3x.png"
+    },
+    {
+      "size": "72x72",
+      "idiom": "ipad",
+      "scale": "1x",
+      "filename": "ipad-72x72-1x.png"
     },
     {
       "size": "72x72",
