+++ test/NativeApp/ios/native_app/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,6 +1,18 @@
 {
   "images": [
     {
+      "size": "40x40",
+      "idiom": "ipad",
+      "scale": "2x",
+      "filename": "ipad-40x40-2x.png"
+    },
+    {
+      "size": "40x40",
+      "idiom": "ipad",
+      "scale": "1x",
+      "filename": "ipad-40x40-1x.png"
+    },
+    {
       "size": "29x29",
       "idiom": "ipad",
       "scale": "2x",
@@ -61,18 +73,6 @@
       "filename": "iphone-29x29-1x.png"
     },
     {
-      "size": "40x40",
-      "idiom": "ipad",
-      "scale": "2x",
-      "filename": "ipad-40x40-2x.png"
-    },
-    {
-      "size": "40x40",
-      "idiom": "ipad",
-      "scale": "1x",
-      "filename": "ipad-40x40-1x.png"
-    },
-    {
       "size": "50x50",
       "idiom": "ipad",
       "scale": "1x",
@@ -91,16 +91,16 @@
       "filename": "ipad-72x72-1x.png"
     },
     {
-      "size": "76x76",
-      "idiom": "ipad",
-      "scale": "1x",
-      "filename": "ipad-76x76-1x.png"
-    },
-    {
       "size": "72x72",
       "idiom": "ipad",
       "scale": "2x",
       "filename": "ipad-72x72-2x.png"
+    },
+    {
+      "size": "76x76",
+      "idiom": "ipad",
+      "scale": "1x",
+      "filename": "ipad-76x76-1x.png"
     },
     {
       "size": "76x76",
