+++ test/ReactNativeIconTest/ios/ReactNativeIconTest/Images.xcassets/AppIcon.appiconset/Contents.json
@@ -1,10 +1,40 @@
 {
   "images": [
     {
+      "size": "50x50",
+      "idiom": "ipad",
+      "scale": "1x",
+      "filename": "ipad-50x50-1x.png"
+    },
+    {
+      "size": "40x40",
+      "idiom": "ipad",
+      "scale": "2x",
+      "filename": "ipad-40x40-2x.png"
+    },
+    {
+      "size": "40x40",
+      "idiom": "ipad",
+      "scale": "1x",
+      "filename": "ipad-40x40-1x.png"
+    },
+    {
+      "size": "29x29",
+      "idiom": "ipad",
+      "scale": "2x",
+      "filename": "ipad-29x29-2x.png"
+    },
+    {
+      "size": "29x29",
+      "idiom": "ipad",
+      "scale": "1x",
+      "filename": "ipad-29x29-1x.png"
+    },
+    {
       "size": "60x60",
       "idiom": "iphone",
-      "scale": "2x",
-      "filename": "iphone-60x60-2x.png"
+      "scale": "3x",
+      "filename": "iphone-60x60-3x.png"
     },
     {
       "size": "57x57",
@@ -43,36 +73,6 @@
       "filename": "iphone-29x29-1x.png"
     },
     {
-      "size": "40x40",
-      "idiom": "ipad",
-      "scale": "1x",
-      "filename": "ipad-40x40-1x.png"
-    },
-    {
-      "size": "50x50",
-      "idiom": "ipad",
-      "scale": "1x",
-      "filename": "ipad-50x50-1x.png"
-    },
-    {
-      "size": "40x40",
-      "idiom": "ipad",
-      "scale": "2x",
-      "filename": "ipad-40x40-2x.png"
-    },
-    {
-      "size": "29x29",
-      "idiom": "ipad",
-      "scale": "2x",
-      "filename": "ipad-29x29-2x.png"
-    },
-    {
-      "size": "29x29",
-      "idiom": "ipad",
-      "scale": "1x",
-      "filename": "ipad-29x29-1x.png"
-    },
-    {
       "size": "50x50",
       "idiom": "ipad",
       "scale": "2x",
@@ -85,10 +85,10 @@
       "filename": "ipad-72x72-1x.png"
     },
     {
-      "size": "60x60",
-      "idiom": "iphone",
-      "scale": "3x",
-      "filename": "iphone-60x60-3x.png"
+      "size": "72x72",
+      "idiom": "ipad",
+      "scale": "2x",
+      "filename": "ipad-72x72-2x.png"
     },
     {
       "size": "76x76",
@@ -97,10 +97,10 @@
       "filename": "ipad-76x76-1x.png"
     },
     {
-      "size": "72x72",
-      "idiom": "ipad",
+      "size": "60x60",
+      "idiom": "iphone",
       "scale": "2x",
-      "filename": "ipad-72x72-2x.png"
+      "filename": "iphone-60x60-2x.png"
     },
     {
       "size": "76x76",
